only register new callback for modal function editor
